BEP-559 - Udates SDK Updates order status for bepado, if it changed in shopware

+++ Library/Bepado/SDK/Gateway/MySQLi.php
@@ -97,6 +97,19 @@ class MySQLi extends Gateway
             }
         }
 
+        return $changes;
+    }
+
+    public function cleanChangesUntil($offset)
+    {
+        $offset = $offset ?: 0;
+        // Float type cast does NOT work here, since the inaccuracy of floating
+        // point representations otherwise omit changes. Yes, this actually
+        // really happens.
+        if (!preg_match('(^[\\d\\.]+$)', $offset)) {
+            throw new \InvalidArgumentException("Offset revision must be a numeric string.");
+        }
+
         // Disable cleanup for the first betas for debuggability and easier re-runs.
         /*$this->connection->query(
             'DELETE FROM
@@ -104,8 +117,6 @@ class MySQLi extends Gateway
             WHERE
                 c_revision <= ' . $offset
         );*/
-
-        return $changes;
     }
 
     private function ensureUtf8($product)
